Improve communication between node-active and pg controller processes.
Rather than guessing from the keeper state file is Postgres should be
running, we use a new Postgres Expected Status file to pass the information
around in between the two processes.

This means that it's possible to control the start and stop of Postgres
directly from anywhere in the code, even halfway through a transition, which
was an unfortunate limit of the previous coding.