Add Postgres logs to our output when Postgres fails to start.
This is huge help to debugging unexpected situations. Also, we go as far as
adding the logfile(s) that have been created after the "startup.log" file
itself. When using this feature, we expect that only one such file exists
and is rather short, and with a FATAL line in there that has a direct hint
to what's going on and how to solve it.