Implement online restart on reload.
When asked to reload, we send SIGHUP to the Postgres process and we just
quit our monitor listener or keeper node-active process, so that our
supervisor process restart them.

This online restart is nice because we are restarting our service from disk
with using an execv() call, which means that's a good way to deploy bug fix
releases without impacting the Postgres service.