Refactor initialization of replicationSource.
It used to be a little under specified about memory management of those
pointers we would keep in the structure, so switch to local memory allocated
in the structure itself and some more memory copying. That way we know when
to de-allocate (on the stack, still). And we can share the initialization
code in a sub-procedure.