Improve pg_autoctl service reloading, and fix ssl.cert_file handling.
In our configuration file we would skip writing the values of the SSL files
when using --ssl-self-signed. The option would still work in so far as we
actually edit the Postgres setup, but then we would fail at reload time.

As we are now depending on more reload-time features, it's time to fix that.