When setting-up a standby, ensure we edit our config with SSL files.
We would not edit the local configuration because we get the Postgres
configuration files from the primary and those are okay already. But we also
need to take into account our own pg_autoctl configuration file, which we
now do in this patch.

For modularity reasons, we can't edit our own configuration file right from
pg_create_self_signed_cert() function, which is sad news, so we have to do
the operation from the caller where we have the KeeperConfig.