Allow disable maintenance in the middle of a failover.
Now that we have multiple standby it might happen that we disable
maintenance on a standby while we are already in the middle of a failover
situation. As we require all nodes to report their LSN position, that's good
news and we might actually be missing that node to report to move forward in
the failover decision making.

Add a transition from CATCHINGUP to REPORT_LSN in the pg_autoctl FSM too, so
that we can actually benefit from this operation. Also, make it so that the
get_primary() primitive returns the current primary even when in the middle
of a failover, adding DRAINING and DEMOTE_TIMEOUT to the list of states that
belong to the primary.