Simplify the ring buffer reading.
We don't actually have to count all the restarts that happened within the
last MaxT seconds, we can just take the oldest one and consider that if this
one happened within the last MaxT seconds then all the more recent ones did
too.

Of course we only do that when we have had more restarts than allowed
already, otherwise the process may be restarted.