Implement a service supervisor for Postgres. (#265)
A service supervisor main job is to call `waitpid()` and take care of the termination of its children processes. It might then restart failed processes. Integrating a service supervision facility allows to later have pg_autoctl be the parent process of Postgres, and also run its main “client” loop in another process: either the monitor listener or the keeper node-active process.

Implement Erlang style restart policy for our services, with the ability to
have permanent, transient, and temporary services managed by a supervisor.
That allows to clean-up the process restart code a lot.

Signals SIGHUP, SIGINT, and SIGTERM are forwarded to children processes. It is also possible to restart only one of the subprocesses, allowing for online restart without impacting the Postgres service.

Co-authored-by: Jelte Fennema <github-tech@jeltef.nl>