WIP: Follow  PostgreSQL's way of handing auth, auth-local and auth-host
When auth option is set, postgres directly sets it to auth-local and auth-host
fields. In fact, it doesn't keep the state of "auth", as can be seen here:
https://github.com/postgres/postgres/blob/5cbfce562f7cd2aab0cdc4694ce298ec3567930e/src/bin/initdb/initdb.c#L3032

This WIP commit aims the same. The external API still gets "auth", in addition
to that "auth-host" and "auth-local". Internally, when "auth" is passed, directly
set it  to "auth-host" and "auth-local".