Remove option --allow-removing-pgdata.
With the recent changes for allowing to register a standby when a PGDATA
directory is found on-disk there is no case where we have an existing PGDATA
that we might be successful registering as a standby only to later remove
from disk so that we can pg_basebackup on-top of it.

That's because we use the system_identifier on the monitor to allow for the
node registration to proceed. When the system_identifier does not match the
other nodes in the same group, the register_node() transaction now fails.