Command pg_autoctl perform failover now waits until failover is done.
We use the same approach as when applying a new setting: listen to
notification messages from the monitor and parse them until we see that we
have reached the expected target state.

In this case it's easy, we now we reached the target state when we see the
notification of a node reaching state primary.