Fix pgautofailover extension upgrade script from 1.2 to 1.3. (#288)
When using Postgres < 12 we can't ALTER an ENUM TYPE within a transaction,
and the extension upgrade script runs in a transaction. Which means we can't
upgrade pgautofailover when using Postgres 12 at the moment.

The fix requires dropping the enum type entirely, keeping the data in its
text form, and casting the data again to the new enum type after it's been
created. It's the long and painful way, but it is required for our support
of Postgres 10 and 11.