Implement different Postgres connection retry policies.
When the keeper loop is running, we should not retry connecting to the
monitor and let the main loop logic take care of potential network split
situations, and take split-brain protective steps.

When running interactive commands, we might want to retry for up to a
minute, as a service to our users.

When provisionning a cluster, some nodes might get ready to connect to the
monitor before that one is live, so it warrants yet another retry policy.

In all cases, we use an Exponential Backoff with Decorrelated Jitter
approach to decide how much time to sleep in between retries, to avoid
hammering the monitor all at once when it's ready / back from the dead.