Per review.
The interactive retry policy is now the default, and the main keeper loop
ensures that we use the specific no-retry policy when contacting the monitor
for the node_active protocol.

Adding to that, the default interactive policy now has a 2 seconds retry
period and can be controlled explicitely thanks to PGCONNECT_TIMEOUT. We
still implement our own retry mechanism within the bounds of
PGCONNECT_TIMEOUT though, rather than the default Postgres one, and force
our PGCONNECT_TIMEOUT to always be 1 second.