Tighten when to accept to perform a failover even more. (#294)
It is only safe to switchover when the current state is primary/secondary. Also, before
reporting that we have reached the goal State of PRIMARY we now make sure that
the standby has reported a LSN that is past the point where we enabled sync rep, to
protect better against data loss in switch overs.