Per review.
Prevent split brain situation during maintenance by preparing the standby
setup as soon as we enter maintenance. A user restarting Postgres would then
have a read-only standby server that has no primary conninfo yet.

Also because we do a clean shutdowwn before promotion of the secondary, we
consider that we won't have to use pg_rewind in this scenario.