Per review, we actually want to pg_rewind from a standby node.
When integrating a standby node to the (possibly new) primary, even when the
node was a standby before, we don't know what happened in the maintenance
window. Users might have had to do some queries on the node, even read-write
queries, and so we need to pg_rewind now.

This simplifies our reasoning with maintenance -> catching up as we always
do a pg_rewind now, rather than only if the node used to be a primary before
reaching maintenance.