Use the normal FSM flow when the primary goes to maintenance.
Trying to use shortcut is not a good idea, then we end up with situations in
the FSM where we don't have a primary, or two of them. Go through the same
steps on the secondary as for a whole failover situation to avoid that
problem, and let the primary in prepare_maintenance long enough to
disambiguate with a failover.