Feature/primary to maintenance (#306)
Allow pg_autoctl enable maintenance on a primary node.

    pg_autoctl enable maintenance --allow-failover

Prevent split brain situation during maintenance by preparing the standby
setup as soon as we enter maintenance. A user restarting Postgres would then
have a read-only standby server.

When integrating a standby node to the (possibly new) primary, even when the
node was a standby before, we don't know what happened in the maintenance
window. Users might have had to do some queries on the node, even read-write
queries, and so we need to pg_rewind now.

This simplifies our reasoning with maintenance -> catching up as we always
do a pg_rewind now, rather than only if the node used to be a primary before
reaching maintenance.

When going down for maintenance on the secondary, make sure we disable sync
replication before we allow users to stop Postgres, in order to avoid
locking writes on the primary, thanks to new state wait_maintenance.

Co-authored-by: Jelte Fennema <github-tech@jeltef.nl>