Reduce how often we exec() pg_controldata.
When Postgres is known to be running we can update our control data
information in the cache from a SQL query rather than doing another fork()
and exec() on the pg_controldata binary and parsing its output.

There are still a few too many calls to pg_controldata after this patch, but
at least we don't do it every second anymore when Postgres is running fine.