Report services version strings in pg_autoctl version --json.
Our new pg_autoctl upgrade procedure will make it possible to have a
supervisor process running with version 1.4 and the node-active process
being restarted from disk with version 1.5. It is then important that we are
able to provide this information to the users.

Currently, the information fits in our output for pg_autoctl version when
using the JSON format, so that's where it is available.

To be able to have that information, we need the services themselves to
publish it. For that, we invent per-service pidfiles and have a new pidfile
format that contains the version string.