Edit the HBA file as soon as a new node is added to the group. (#311)
We don't have to wait until a failover happens to install the HBA rules we
need. Also installing those at node creation would help users review and
edit them, if needed.

Also, share the otherNodesArray cache with the other procedures that are
needing it, namely the replication slot maintenance procedures. Finally,
refactor the node_active code to make it easier to maintain in the long run.

pg_basebackup replaces our pg_hba.conf with the one from the upstream
server, and whatever we think we edited in there might have been lost. Add a
parameter that allows keeper_refresh_other_nodes() to ensure HBA rules are
present for all the other nodes, no matter what we had in the previous loop
run.