Only use replication slots on the standby nodes for good Postgres version.
A bug in Postgres prevents WAL recycling on standby nodes when we use the
function pg_replication_slot_advance() to maintain their LSN position. In
this patch we work around the bug to avoid it by just not installing our
replication slots on the standby nodes when we're running a Postgres version
with the bug. That's 11.0 to 11.8 included, and 12.0 to 12.3 included.

Fixes #283.