Additional cleaning-up for PATH computations.
When PGDATA points to an existing directory, use realpath(3) on a relative
pathname rather than playing with CWD tricks. In some cases we have a race
condition when computing the Postgres Expected Status file and then we fail
to start Postgres. When ensuring that we always compute the path the same
way, we seem to be solving that race condition around the creation of the
PGDATA directory.

See #329 again.