Fix test environment detection.
In current version of pg_autoctl the sub-processes are running with the
PG_AUTOCTL_DEBUG environment variable set. That's because we invoke the
sub-processes with a debug level command: pg_autoctl do service ...

As we still want to test replication slots handling in the test suite,
detect that we are in that context by using both PG_AUTOCTL_DEBUG and also
PG_REGRESS_SOCK_DIR, which is only used there.