Improve race-condition protection in the code some more.
The problem is that we compute the Postgres expected status pathname from
the user-given PGDATA even when PGDATA has been created on-disk. Ensure that
we always use the normalized path as soon as PGDATA exists. That's the only
way to guarantee that all the processes are going to use the same path.