Remove two states: WAIT_FORWARD and WAIT_CASCADE.
Those states were needed to orchestrate the failover and give the
opportunity for the new upstream node (the standby with the most advanced
LSN) to prepare its HBA config. Nowadays the HBA permissions are edited
eagerly, at node registration, so we don't have to implement that dance
anymore.

We need a new way to find the node with the most advanced LSN though, where
before we would just find the other node with the state WAIT_CASCADE. Turns
out finding such a node is as easy as ORDER BY reportedlsn DESC LIMIT 1.