Default number_sync_standbys is now zero, that's easier to understand.
With a number_sync_standbys of zero, we still use synchronous replication to
ensure a smooth failover when that needs to happen. Unless the node is set
with replication-quorum to false.

When we add a second standby node, the third Postgres node in a group, then
we automatically switch the number_sync_standbys to 1 so that we have the
expected general behavior of allowing to lose a standby node, and blocking
writes on the primary when losing a second standby node. When using
replication-quorum true on all nodes, that is.