Rename --nodename to --hostname, and nodename to nodehost. (#273)
* Rename internal variable "hostname" to "nodenameCandidate".

In discover_nodename() our usage of variable names can be confusing and
conflicts with the intended changes done next in this PR commit series.

* Apply script to replace nodename with hostname.

With the following script we edit the code and replace nodename with
hostname, in preparation for adding an actual node name which is not the
technical hostname used in the connection string.

    sed -i 's/NodeName/NodeHost/g'   src/monitor/{*.{c,h,sql},expected/*.out,sql/monitor.sql}
    sed -i 's/nodeName/nodeHost/g'   src/monitor/{*.{c,h,sql},expected/*.out,sql/monitor.sql}
    sed -i 's/nodename/nodehost/g'   src/monitor/{*.{c,h,sql},expected/*.out,sql/monitor.sql}
    sed -i 's/NODENAME/NODEHOST/g'   src/monitor/{*.{c,h,sql},expected/*.out,sql/monitor.sql}
    sed -i 's/NODE_NAME/NODE_HOST/g' src/monitor/{*.{c,h,sql},expected/*.out,sql/monitor.sql}
    sed -i 's/nodeName/hostName/g'   tests/*.py src/bin/pg_autoctl/*.{c,h}
    sed -i 's/nodename/hostname/g'   tests/*.py src/bin/pg_autoctl/*.{c,h}
    sed -i 's/NodeName/HostName/g'   tests/*.py src/bin/pg_autoctl/*.{c,h}
    sed -i 's/NODENAME/HOSTNAME/g'   tests/*.py src/bin/pg_autoctl/*.{c,h}
    sed -i 's/NODE_NAME/HOST_NAME/g' tests/*.py src/bin/pg_autoctl/*.{c,h}

* Refactor nodename to hostname in the documentation.

* Implement compatibility option pg_autoctl.nodename.

This allows reading the old setting name in the configuration file and when
we write the file again, we use the new setting name. That way the
transition from nodename to hostname is all automated at pg_autoctl upgrade.