Fix the recent refactoring for building the candidate list.
In the refactoring we changed the counters that we maintain and so some of
the logic around. The place were to assign DRAINING to the primary was not
well defined enough and failing to assign DRAINING had dire consequences.

Also, the SQL query embedded in monitor.c to fet the most advanced standby
required several round of fixes, because of the evolution of the extension
schema on the monitor. It's best to move that query to the extension itself
so that we maintain it at the right place.