Allow remove_node() even on a primary when having multiple standby nodes.
It's important for operators to be able to remove a node whenever that's
needed. Either as a last resort when the FSM is otherwise stucked, or when
the node is actually never going to be available again and we can't decide
if that's the case or a network split.

When removing a primary node on a single standby system, we simply
transition the only remaining node to the SINGLE state. That's easy, and the
FSM both on the monitor and on the keeper know how to proceed already.

When removing a primary node on a multiple standby syste, we need to trigger
a standby election mechanism for the failover, and the code needs to be
ready to accept that there's no primaryNode (it's NULL).