Fix secondary -> catching-up to check if there's a new primary.
This transition is only used when a secondary node is known unhealthy from
the monitor. If the node is in a network split, it will fail to connect to
the monitor and only will get the new CATCHINGUP assignement when it can
connect again, and meanwhile a failover might have happened.

So let's have the standby node renew its upstream/primary setup when
implementing the secondary to catchingup transition.