Actually check the primary_conninfo we setup.
Use pg_receivewal to connect to the primary, with an --endpos that we
already have received so that this does nothing on the server. All it does
is checking that it can connect to the primary with the connection string
and replication slot given.

This should allow to fail early when the HBA on the other node has not been
made ready yet for reasons, and retry in the next iteration, or actually
fail with an pg_autoctl error message rather than a Postgres log.