Exit early with EXIT_CODE_QUIT in pg_autoctl create when possible. (#349)
When we return false from keeper_pg_init_and_register() the initialisation
process is then exited with an error code, and the supervisor is going to
try again to run the TRANSIENT service successfully. This makes sense when
we failed to contact to monitor or had another kind of intermitent failure.

When the problem is that it's all good and you should use --run or the
pg_autoctl run command, well, early exit with success code is better:
arguably we did our job.