Refrain from using pg_receivewal --synchronous.
We are using pg_receivewal to test for the primary_conninfo connection
string. We don't actually want to receive WAL files, and we certainly don't
want to have the client connections wait on us when we do that.

Also our target LSN is 0/1 which should never require the connection to
actually download any WAL bits.