Add more multi-standby maintenance tests, and some fixes.
Turns out the JOIN_SECONDARY state was being used in more cases than it
should have been. In particular, we want the primary to be in the PRIMARY
state as soon as there's at least one SECONDARY node that we can failover
to.

Once that decision is clear, we need to enable the transition from
REPORT_LSN directly to SECONDARY. Also, some others group FSM changes can be
simplified and dealt with in the main code area, without entering the loops
to find and select a candidate for failover.