Be smart about pg_replication_slot_advance().
Reading the source code for Postgres (src/backend/replication/slotfuncs.c
for the function pg_replication_slot_advance) we see that the minimum LSN we
can advance too on a physical replication slot is the restart_lsn.

Rather than trying to be smart about error handling when the monitor or
another standby in the system is lagging too much, just refrain from calling
the pg_replication_slot_advance function when the target LSN is before the
restart_lsn of the slot.