Allow multiple standbys in a Postgres group. (#326)
Allow to have more than two nodes per Postgres group, and orchestrate failover
with those nodes when needed. Note that in particular:

  - when there's only two nodes in a group, number_sync_standby is zero
  - when adding a third node in a group, number_sync_standby is set to one

With number_sync_standby set to one, we can only put one of the two standby nodes to
maintenance at a time, and in the unfortunate case where both the standby nodes would
become unhealthy, then we keep synchronous replication in place.

To have the magic behaviour of disabling synchronous replication in the event of losing
two many standby nodes, number_sync_standbys must be set to zero manually.

The synchronous_standby_names setting is how we implement our policy of
number_sync_standbys, candidate priorities, and replication quorum. The 
monitor now computes the value in a function that is called from pg_autoctl.

The orchestration of a failover is now more complex as it depends on more
than two nodes. In particular, the nodes are now "anonymous", we need to
cater with more things being dynamic and discovered at run-time, where
before we had just two nodes. 

We introduce new states REPORT_LSN, FAST_FORWARD, and JOIN_SECONDARY to
help with orchestrating multiple standby failover operations, where we need to select
a candidate node for failover based on candidate priorities and LSN positions.

Co-authored-by: Önder Kalacı <onder@citusdata.com>