Refrain from useless verifications of pg_controldata properties. (#342)
We don't need to check all our Postgres cluster invariants when Postgres is
not running. We're going to check for those as soon as Postgres is running
again anyway.

As seen in #341 too many checks for our invariants may block progress in the
FSM on the client side in edge cases.