Fix dropping the old-primary node.
When we remove the old primary, its reported state is still that of a
primary. We might have moved on to a new primary already though, and in that
case we want to process the old-primary the same way we would a secondary.

To fix, consider the goalState rather than the reportedState, and filter out
the DRAINING, DEMOTE, and DEMOTE_TIMEOUT as meaning that we have a primary,
by testing with CanTakeWritesInState() rather than StateBelongsToPrimary().