In some cases, we still need to call pg_controldata. (#357)
As we renew the keeper->postgres.pgSetup cache on each round, we need to
manually make sure to re-install the current values we have in cache when
Postgres is not running. That said, we might also reach that code path at
pg_autoctl start-up when Postgres is not yet running.