Make code blocks show up
This behavior differed between my local version of sphinx and Read The
Docs. Sphinx detects the indent level of a code block by that of the
first line. The table headings were more indented than the table
separator, so it confused the parser. As a hack I include a small
character (.) at the correct level of indentation on the first line.