Cleanup/show state (#336)
* Clean-up code that's not used anymore.

  - remove the pg_autoctl show nodes command, not needed anymore
  - remove pg_autoctl show synchronous-standby-names

The second one was left there as a synonym to `pg_autoctl show
standby-names` and this commits makes a choice between the two spellings.

* Review pg_autoctl show state to include health and LSN.

That's way more useful to understand the current state of a group.

* Send notifications in JSON format.

Now that we have a JSON parser on the keeper side, it makes things simpler,
in particular now that we need to add the node name field. Also it makes it
trivial to now write another consumer for pg_auto_failover notifications.

* pg_autoctl show state --local

Take local data that we have and publish them in the same format as the data
we would usually get from the monitor.