Fix tutorial typos, and make sure all code blocks are displayed (#368)
* Make code blocks show up

This behavior differed between my local version of sphinx and Read The
Docs. Sphinx detects the indent level of a code block by that of the
first line. The table headings were more indented than the table
separator, so it confused the parser. As a hack I include a small
character (.) at the correct level of indentation on the first line.

* hba uses hostssl now

* Typo

* Use bash syntax highlighting

If no syntax format is specified, I think it can cause RTD to also
omit the code blocks