Fail fast if pg_stat_statements is not installed
pg_stat_statements is required on the primary and secondary nodes. If
the extension is not installed in the server we will fail fast and exit
while running pg_autoctl create postgres. This is typical if we are
operating with a dev build of the Postgres server.

Otherwise, we fail much later, while "pg_autoctl create postgres .."
does initdb.
Primary's startup.log:
2020-08-25 15:52:07.680 PDT [47143] FATAL:  could not access file "pg_stat_statements": No such file or directory
2020-08-25 15:52:07.680 PDT [47143] LOG:  database system is shut down