Set password_encryption if auth is password based
This automatically sets the password_encryption GUC while performing a
create {postgres|monitor} if the supplied --auth method is password
based (one of ["md5", "password", "scram-sha-256"]).

This saves the cluster operator from having to set this GUC and reload
the server configuration on each node in case the --auth option is
either {password, scram-sha-256}. (The operator didn't have to set this
GUC for --auth md5 as the default value of password_encryption is md5)

Since the "password" authentication supports only the "md5" and
"scram-sha-256" password encryption settings, if it is supplied we
default the GUC to "scram-sha-256" as it is more secure.