Review the process to register existing nodes to a new monitor.
It's not as bad as I though it would be yesterday. After playing with it
locally where the pg_autoctl create commands are in the history, it's
actually pretty easy to stop and destroy the monitor, install a new one, and
then re-attach every node.

This requires a Postgres restart though.