At startup check if sysidentifier is zero on the monitor and reset it then.
In the upgrade script from 1.3 to 1.4 we set the system identifier to zero
on every node, because we don't have the actual value, and NULL is forbidden
once past the initialisation of nodes in 1.4.

At pg_autoctl run startup we can upload our node system identifier and set
it for the whole group when the current one is zero. We do that at every
startup, because we don't track if we just got updated.