Improve monitor listener main loop.
The new code consumes less CPU and should react quicker to notifications
being available on the monitor's connection. Also, we integrate checks for
Postgres to be running in the main listener loop, retrying to connect to the
database when it's not ready yet, rather than bailing out to the main
supervision process.