Fix a corner case which leaves a node in JOIN_PRIMARY state.
When removing a standby node just after creating it, and before running its
pg_autoctl service, then it's still in the CATCHINGUP state, and the primary
is waiting in the JOIN_PRIMARY state. At removal of the standby, the primary
should feel free to set itself back to PRIMARY again.