Allow registering multiple standbys concurrently (#395)
* Allow concurrent standby nodes registrations.

Now that we edit the HBA file eagerly and for all known nodes, we can move
the responsibility of handling the race conditions to the the client-side.
Which already knows how to retry a failed transition.

* Check that the target replication slot exists before base backup.

There's a bug in pg_basebackup where it only realises that the target
replication slot does not exists after having finished to copy all the files
from its target server. To avoid having to copy a possibly large PGDATA
over, we manually check the condition before allowing pg_basebackup to get
to work.