When secondary, don't ensure current state before a transition. (#402)
Getting out of the secondary state means we should not ensure the secondary
state anymore, it's already too late. We had cases of Postgres being
restarted because of that, and if that happens when we don't have a recovery
setup in place, then Postgres restarts as a primary. Then it can't be a
secondary again anymore, WAL are "corrupted".