Fix some problems in make cluster
My main problems were with the fact that nodes were spawned at the same
time and the node with name `node_2` was not the node with directory
`node2`. This is fixed by doing three things:
1. Adding `--name` to the create commands
2. Make `show state` order by name instead of nodeid, so `node2` would
   actually be above `node3` in the table
3. Fix the `do wait` commands to use `PG_AUTOCTL_DEBUG`.

My other issue was that I could not use `--pgdata node2` in the
interactive shell. This had two reasons:
1. It did not actually `cd` to the `TMUX_ROOT` directory
2. The `XDG` paths were relative, so once I `cd`ed there the XDG paths
   were suddenly wrong.