Force closing the connection used to check replication slots.
We connect to the primary using the pgautofailover_replicator username to
check if the replication slot has been created already on the primary, but
in that case we didn't have another use for the connection, no clean-up
happens later.

Close the connection as soon as we have the information we are looking for.
We reconnect each time anyway.