We could reach join_secondary while the primary is not stable.
And we need to make progress to allow the primary to reach a stable
situation, it might be wait_primary/primary because another secondary has
done joining again, but is not a candidate for failover, thus we can't move
towards primary. If now the other secondary nodes can't join in this
intermediate state, the machine is clogged.