Track class 08 (connection exception) as PG_CONNECTION_BAD.
This allows to have a better error handling mechanism, in particular when we
fail to get the monitor pgautofailover extension version because we lost the
connection, at monitor startup. We need to be able to track that problem as
a PG_CONNECTION_BAD problem so that we know how to retry rather than exit
early.