Implement HBA hostname DNS checks, and warn users when needed. (#458)
Postgres HBA implementation with hostname is pretty picky, in order to be
secure enough. The reverse DNS lookup requirements are often missed, and
sometimes hard or impossible to achieve (Cloud VM environments).

As more users get confused with their hostname based setup, we should
provide more guidance when that's possible.