Get rid of the APPLY_SETTINGS state and transitions.
In case an upgrade should happen while the primary is in APPLY_SETTINGS, we
still keep some level of compatibility with the state: we still know how to
move the FSM forward in that case.

That said, we don't assign APPLY_SETTINGS anymore and apply cache
invalidation to the synchronous_standby_names settings in the same way we
have done already for the otherNodes list and the HBA editing.

Finally, when using pg_autoctl set node|formation ... we return directly to
the user, we don't ensure the setting made its way to the primary node
already when the command is finished.