Discover the monitor default hostname with gethostname(3).
When we have neither --monitor nor --hostname, use gethostname(3) (same as
hostname -I on linux) and use that when a DNS lookup resolves to at least
one IP address that is also present on the machine.

Otherwise, default to contacting the google DNS service to find the default
IP address to use that has a chance of being available to other nodes.