Review memory/malloc errors and potential leaks. (#478)
We avoid using malloc in most of the code base, and still require it when
reading a file from disk. Some places in the code didn't get the memo that
read_file() and read_file_if_exists() leak memory, and the caller needs to
call free().

Also, because we only have a handful of malloc() failed error messages, it's
a good thing to always use the same error message format. And this format
should not require memory allocation of course. I believe %m is okay with
that.