Fix pg_autoctl do show hostname.
When we can't use the result of gethostname(3) then we need to continue with
the full discovery dance, the previous code would wrongly use
DEFAULT_INTERFACE_LOOKUP_SERVICE_NAME (which is 8.8.8.8).

Fortunately this bug is not replicated in the pg_autoctl create monitor code
path, so it's only a bug in the helper debug function.