Review APPLY_SETTINGS transitions in cases including node removal.
When removing a node we might end up with a race condition if the primary in
parallel is already in a transition, or in any other case when
ProceedGroupState on the primary doesn't change the assigned state.

In that case we now force a new APPLY_SETTINGS state to be assigned, so that
the primary has a chance of fetching the (new) current synchronous standby
names settings.

Also, it might be that a node has been dropped while reaching APPLY_SETTINGS
itself, so in the transition from APPLY_SETTINGS to PRIMARY we also
implement fsm_apply_settings, to protect against race conditions.

Any transition to PRIMARY now also calls fsm_apply_settings.