test_multi_standbys: add failing unit test
Rewrite test_014 to exercise a "simultaneous" add/drop situation, which
leads to a hung FSM. This does not fix the problem, just expose it.

One additional issue: the wait_until_state(join_primary) call is itself
racing against the primary, which might have already transitioned out of
that state by the time we reach the call. I'm not sure how to fix that
without a lot more plumbing.

Based on code originally authored with rheaton.

Co-authored-by: Rachel Heaton <rheaton@pivotal.io>