Fix when we switch synchronous_standby_names to '*'.
Because we assign both SECONDARY and PRIMARY in the same monitor
ProceedGroupState call, we might end up with the primary fetching its new
synchronous_standby_names from the monitor before when the secondary has
made it to its new state.

With the previous coding, the primary would then retrieve '', meaning sync
replication is disabled, when what's needed is '*'.

We add some needed test coverage from the situation and fix it in the
monitor by returning '*' as soon as the other node is assigned SECONDARY
rather than only after it has reported the state.