Fix/non default formation (#489)
* Refrain from setting the default formation on most commands.

When we are supposed to read a configuration file, rather than creating one
for the first time, avoid setting a default formation in our command line
parsing structure. It's impossible to then distinguish between the user
saying --formation default to override what's in the configuration with the
case where we just want to use whatever is the default.

* Fix default handling of --formation.

Any command that needs to operate on a pre-existing node must read the
node's formation from its configuration file, rather than default the maybe
missing command line argument --formation to "default", which obviously only
works when the node happens to have been created in the "default" formation.