Fix syncrhonous_standby_names when disabling maintenance.
When a node gets out of maintenance and back in the group, we might have to
change the syncrhonous_standby_names on the primary to include the node
again. For that, if the primary node is in a stable PRIMARY state, thanks to
other standby nodes being available, we need to assign APPLY_SETTINGS to the
current primary.