Introduce unit testing for --disable-monitor, and fix many places.
It turns out that many places in the code assumed that we could contact a
monitor to grab the list of other nodes, get the current primary node, or
get the most advanced standby. Use the keeper->otherNodes structure that we
load from file when using --disable-monitor.

Also, as it's usually the monitor's responsibility to assign a nodeId to
every node that is registered, introduce the new --node-id option that is
mandatory when --disable-monitor is used.

The JSON array file used in `pg_autoctl do fsm nodes set ...` can contain
the full list of nodes, it is now filtered so that the keeper->otherNodes
list only contains the other nodes.