Fix/debian stats temp directory (#504)
* Add stats_temp_directory to the list of GUCs to comment on debian.

When creating a Postgres cluster in debian, the stats_temp_directory is
rewrittent to a debian specific place, created by the debian init scripts at
boot time. As we unregister the debian Postgres clusters to the debian init
scripts, we can't rely on their setting and tricks for stats_temp_directory.

* The Dockerfile is now using Postgres 12, adjust PGVERSION too.

This fixes tests/test_debian_clusters.