Fix --disable-monitor to work with the parenting of Postgres model. (#494)
* Fix --disable-monitor to work with the parenting of Postgres model.

In the current version the pg_autoctl process is the parent process of the
Postgres main process. Except when using --disable-monitor, which makes no
sense, so here we fix that.

It is now possible to control and run a Postgres process under a pg_autoctl
process without using a monitor, as in the following example:

  $ pg_autoctl create postgres --auth trust --no-ssl --disable-monitor --run
  $ PG_AUTOCTL_DEBUG=1 pg_autoctl do fsm assign single

The `pg_autoctl do fsm assign` command now edits the keeper state file, and
the FSM transition that is necessary is then implemented in the keeper loop
as usual.

* Introduce a nodes array file (JSON) for --disable-monitor use case.

This allows users to fill-in the list of nodes in a cluster, so that
transitions that create replication slots and HBA entries can work without a
monitor.

New commands are introduced to manage (parse/install) the nodes array:

  pg_autoctl do fsm nodes
    get  Get the list of nodes from file (see --disable-monitor)
    set  Set the list of nodes to file (see --disable-monitor)

* Move test_monitor_disabled to the "multi" series.

It seems otherwise we hit a threshold where Travis can't run the whole test
suite anymore?

* Remove unfinished test 14 in test_multi_standbys.

This test is not finished and not stable. Should be added back later in its
own PR when ready.