[POC] Add rudimentary support to pgbouncer
TL;DR Assuming that pgbouncer is installed, allow pg_autoctl to control it.

This commit adds the 'pgbouncer' subcommand to 'create'. So, when a user issues:

$ pg_autoctl create pgbouncer [options]

then, pg_autoctl will control the pgbouncer instance that allows users to
connect to
a primary node. When the state of the primary node changes, then pg_autoctl
instructs pgbouncer to pause until there exists a stable primary again. At that
point, pg_autoctl will update the pgbouncer's configuration accordingly to
follow the
new primary and unpause it.

At this point, pg_autoctl runs in the foreground and spawns a child that runs
the pgbouncer command. No effort to downgrade child privileges, redirect
streams, etc, is being made. However, a small effort is made to make certain
that
the child process is running, otherwise pg_autoctl exits.

In order to point pgbouncer to the primary, pg_autoctl takes ownership of the
[databases] section of the pgbouncer's configuration. It constructs the
necessary
connection strings on startup and every time there is a state change.

There are no efforts made to make certain that the user-supplied configuration
is valid.

The current set of options are:
    --config    Path to pgbouncer configuration
    --monitor   URL to connect to pg_auto_failover monitor

No tests are added at this point.