Use PG_CONFIG has a hint to choose one of multiple Postgres versions.
When more than one Postgres version is found in $PATH a classic way to
target a specific one, for Postgres contributors and extension authors,
consists of using PG_CONFIG.

With this PR we make it so that pg_autoctl also follows PG_CONFIG and use
the pg_ctl PATH in the same directory as the pg_config setup that way in the
environment.

In passing, cleanup the pg_ctl_version() API to stop leaking memory out of
its local stack/heap.