Code cleaning, refactoring, and review of the log messages.
We get rid of the dynamic memory allocation scheme of the search_path()
function to prefer a now classic strategy of stack allocation in a
structure.

Also make a better choice of when to search for pg_config in the PATH in
situations where there are either 0 or > 1 pg_ctl entries found in PATH.

Finally, add a new command to try things out:

  PG_AUTOCTL_DEBUG=1 pg_autoctl do pgsetup pg_ctl --pgdata /tmp/m

The PGDATA directory does not need to exists.