Allow a DEMOTED primary to come back to a new wait/join/primary.
It used to be that when showing back up a demoted node could only see a
wait_primary node, but that's not the case anymore. Another standby being
available could allow the new primary to be already a PRIMARY, or maybe in
the process of becoming one, in the JOIN_PRIMARY state.