Per review, refrain from assigning primary when we don't have a candidate.
We want to reach PRIMARY/SECONDARY only when we have a known candidate for
failover, which means a node that is both healthy and with a non-zero
candidate priority.

Otherwise we have a loop of failures to reach the Target LSN from any
secondary in the synchronous_standby_names list, where this list should stay
empty given the known current situation.